eatgrediant_ai_api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-aiplatform==1.65.0
python-multipart==0.0.6
Pillow==10.4.0
google-auth==2.23.4
google-cloud-storage==2.10.0
